feat(l1): remove revm from workspace (#4686)
**Motivation**
Remove rem as a workspace dependency
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Remove `revm` feature from all workspace crates
* Remove revm backend from `ethrex-vm` crate
* Remove `tooling/ef_tests/state` and
`crates/vm/levm/bench/revm_comparison` from workspace (They still retain
revm as a dependency)
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number
