+++ crates/l2/prover/Cargo.toml
@@ -38,7 +38,7 @@ ethrex-sdk.workspace = true
 
 guest_program = { path = "./src/guest_program", default-features = false }
 
-risc0-zkvm = { version = "=3.0.3", optional = true }
+risc0-zkvm = { version = "=3.0.3", optional = true, features = ["getrandom"] }
 risc0-zkp = { version = "=3.0.2", optional = true }
 
 sp1-sdk = { version = "=5.0.8", optional = true }
