+++ crates/l2/prover/src/guest_program/Cargo.toml
@@ -24,7 +24,7 @@ ethrex-l2-common = { path = "../../../common", default-features = false }
 [build-dependencies]
 hex.workspace = true
 risc0-build = { version = "=3.0.3", optional = true }
-risc0-zkvm = { version = "=3.0.3", optional = true }
+risc0-zkvm = { version = "=3.0.3", optional = true, features = ["getrandom"]  }
 sp1-build = { version = "=5.0.8", optional = true }
 sp1-sdk = { version = "=5.0.8", optional = true }
