+++ crates/l2/prover/src/guest_program/src/risc0/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2024"
 
 [dependencies]
 risc0-zkvm = { version = "=3.0.3", default-features = false, features = [
-    "std",
+    "std", "getrandom"
 ] }
 risc0-zkvm-platform = { version = "=2.2.1", default-features = false, features = [
     "sys-getenv",
