+++ crates/l2/tee/quote-gen/Cargo.toml
@@ -21,7 +21,7 @@ guest_program = { path = "../../prover/src/guest_program", default-features = fa
 ethrex-l2 = { path = "../..", default-features = false }
 ethrex-l2-common = { path = "../../common", default-features = false }
 
-secp256k1 = { version = "0.29.1", default-features = false, features = [
+secp256k1 = { version = "0.30.0", default-features = false, features = [
   "global-context",
   "recovery",
   "rand",
