+++ crates/networking/p2p/rlpx/l2/l2_connection.rs
@@ -219,11 +219,12 @@ pub(crate) async fn send_new_block(established: &mut Established) -> Result<(), 
                             &l2_state.committer_key,
                         )
                         .serialize_compact();
-                    let recovery_id: u8 = recovery_id.to_i32().try_into().map_err(|e| {
-                        RLPxError::InternalError(format!(
-                            "Failed to convert recovery id to u8: {e}. This is a bug."
-                        ))
-                    })?;
+                    let recovery_id: u8 =
+                        Into::<i32>::into(recovery_id).try_into().map_err(|e| {
+                            RLPxError::InternalError(format!(
+                                "Failed to convert recovery id to u8: {e}. This is a bug."
+                            ))
+                        })?;
                     let mut sig = [0u8; 65];
                     sig[..64].copy_from_slice(&signature);
                     sig[64] = recovery_id;
