+++ crates/networking/p2p/rlpx/l2/messages.rs
@@ -65,7 +65,7 @@ impl BatchSealed {
         let (recovery_id, signature) = secp256k1::SECP256K1
             .sign_ecdsa_recoverable(&SecpMessage::from_digest(hash.into()), secret_key)
             .serialize_compact();
-        let recovery_id: u8 = recovery_id.to_i32().try_into().map_err(|e| {
+        let recovery_id: u8 = Into::<i32>::into(recovery_id).try_into().map_err(|e| {
             RLPxError::InternalError(format!(
                 "Failed to convert recovery id to u8: {e}. This is a bug."
             ))
