+++ crates/networking/rpc/eth/transaction.rs
@@ -12,9 +12,7 @@ use crate::{
 use ethrex_blockchain::{Blockchain, vm::StoreVmDatabase};
 use ethrex_common::{
     H256, U256,
-    types::{
-        AccessListEntry, BlockHash, BlockHeader, BlockNumber, Fork, GenericTransaction, TxKind,
-    },
+    types::{AccessListEntry, BlockHash, BlockHeader, BlockNumber, GenericTransaction, TxKind},
 };
 
 use ethrex_rlp::encode::RLPEncode;
@@ -108,15 +106,12 @@ impl RpcHandler for CallRequest {
             // Block not found
             _ => return Ok(Value::Null),
         };
-        let chain_config = context.storage.get_chain_config()?;
-        let fork = chain_config.get_fork(header.timestamp);
         // Run transaction
         let result = simulate_tx(
             &self.transaction,
             &header,
             context.storage,
             context.blockchain,
-            fork,
         )?;
         serde_json::to_value(format!("0x{:#x}", result.output()))
             .map_err(|error| RpcErr::Internal(error.to_string()))
@@ -353,12 +348,9 @@ impl RpcHandler for CreateAccessListRequest {
 
         let vm_db = StoreVmDatabase::new(context.storage.clone(), header.hash());
         let mut vm = context.blockchain.new_evm(vm_db)?;
-        let chain_config = context.storage.get_chain_config()?;
-        let fork = chain_config.get_fork(header.timestamp);
 
         // Run transaction and obtain access list
-        let (gas_used, access_list, error) =
-            vm.create_access_list(&self.transaction, &header, fork)?;
+        let (gas_used, access_list, error) = vm.create_access_list(&self.transaction, &header)?;
         let result = AccessListResult {
             access_list: access_list
                 .into_iter()
@@ -465,9 +457,6 @@ impl RpcHandler for EstimateGasRequest {
             }
         };
 
-        let chain_config = storage.get_chain_config()?;
-        let fork = chain_config.get_fork(block_header.timestamp);
-
         // If the transaction is a plain value transfer, short circuit estimation.
         if let TxKind::Call(address) = transaction.to {
             let account_info = storage
@@ -482,7 +471,6 @@ impl RpcHandler for EstimateGasRequest {
                     &block_header,
                     storage.clone(),
                     blockchain.clone(),
-                    fork,
                 );
                 if let Ok(ExecutionResult::Success { .. }) = result {
                     return serde_json::to_value(format!("{TRANSACTION_GAS:#x}"))
@@ -515,7 +503,6 @@ impl RpcHandler for EstimateGasRequest {
             &block_header,
             storage.clone(),
             blockchain.clone(),
-            fork,
         )?;
 
         let gas_used = result.gas_used();
@@ -544,7 +531,6 @@ impl RpcHandler for EstimateGasRequest {
                 &block_header,
                 storage.clone(),
                 blockchain.clone(),
-                fork,
             );
             if let Ok(ExecutionResult::Success { .. }) = result {
                 highest_gas_limit = middle_gas_limit;
@@ -580,12 +566,11 @@ fn simulate_tx(
     block_header: &BlockHeader,
     storage: Store,
     blockchain: Arc<Blockchain>,
-    fork: Fork,
 ) -> Result<ExecutionResult, RpcErr> {
     let vm_db = StoreVmDatabase::new(storage.clone(), block_header.hash());
     let mut vm = blockchain.new_evm(vm_db)?;
 
-    match vm.simulate_tx_from_generic(transaction, block_header, fork)? {
+    match vm.simulate_tx_from_generic(transaction, block_header)? {
         ExecutionResult::Revert {
             gas_used: _,
             output,
