+++ crates/vm/Cargo.toml
@@ -11,8 +11,6 @@ ethrex-levm = { path = "./levm", default-features = false }
 ethrex-trie.workspace = true
 ethrex-rlp.workspace = true
 
-# These dependencies must be kept up to date with the corresponding revm version, otherwise errors may pop up because of trait implementation mismatches
-
 derive_more = { version = "1.0.0", features = ["full"] }
 bytes.workspace = true
 thiserror.workspace = true
@@ -29,18 +27,6 @@ dyn-clone = "1.0"
 
 ethereum-types.workspace = true
 
-revm = { version = "19.0.0", features = [
-    "serde",
-    "std",
-    "serde-json",
-    "optional_no_base_fee",
-    "optional_block_gas_limit",
-], default-features = false, optional = true }
-revm-inspectors = { version = "0.15.0", optional = true }
-revm-primitives = { version = "15.2.0", features = [
-    "std",
-], default-features = false, optional = true }
-
 [lib]
 path = "./lib.rs"
 
@@ -48,13 +34,6 @@ path = "./lib.rs"
 default = []
 c-kzg = ["ethrex-levm/c-kzg", "ethrex-common/c-kzg"]
 debug = ["ethrex-levm/debug"]
-revm = [
-    "dep:revm",
-    "dep:revm-inspectors",
-    "dep:revm-primitives",
-    "revm/blst",
-    "revm/c-kzg",
-]
 
 [lints.clippy]
 unwrap_used = "deny"
