+++ crates/vm/levm/bench/revm_comparison/Cargo.toml
@@ -1,9 +1,10 @@
 [package]
 name = "revm_comparison"
-version.workspace = true
-edition.workspace = true
-authors.workspace = true
-documentation.workspace = true
+version = "0.1.0"
+edition = "2024"
+authors = ["LambdaClass"]
+
+[workspace]
 
 [lib]
 name = "revm_comparison"
@@ -11,12 +12,12 @@ path = "src/lib.rs"
 
 [dependencies]
 ethrex-levm = { path = "../../" }
-ethrex-vm.workspace = true
-ethrex-common.workspace = true
-ethrex-storage.workspace = true
-ethrex-blockchain.workspace = true
-hex.workspace = true
-bytes.workspace = true
+ethrex-vm = { path = "../../.." }
+ethrex-common = { path = "../../../../common" }
+ethrex-storage = { path = "../../../../storage" }
+ethrex-blockchain = { path = "../../../../blockchain" }
+hex = "0.4.3"
+bytes = { version = "1.6.0", features = ["serde"] }
 
 revm = "9.0.0"
 sha3 = "0.10.8"
