+++ tooling/ef_tests/blockchain/Cargo.toml
@@ -29,7 +29,6 @@ path = "./lib.rs"
 [features]
 default = ["c-kzg"]
 c-kzg = ["ethrex-blockchain/c-kzg"]
-revm = ["ethrex-vm/revm"]
 sp1 = ["guest_program/sp1", "ethrex-prover/sp1"]
 stateless = []
